Build Tools Task 1 completed

+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -3,6 +3,6 @@ package com.epam.utils;
 public class StringUtils {
     public static boolean isPositiveNumber(String str) {
          //here magic will happen
-        return org.apache.commons.lang3.math.NumberUtils.toInt(str) > 0;
+        return org.apache.commons.lang3.math.NumberUtils.toDouble(str) > 0;
     }
 }
